chore: deduplicate ws message
