+++ libs/client-api/src/ws/msg_queue.rs
@@ -53,7 +53,7 @@ impl AggregateMessageQueue {
     let maximum_payload_size = self.maximum_payload_size;
     let weak_queue = Arc::downgrade(&self.queue);
     let weak_seen_ids = Arc::downgrade(&self.seen_ids);
-    let interval_duration = Duration::from_secs(1);
+    let interval_duration = Duration::from_millis(500);
     let mut next_tick = Instant::now() + interval_duration;
     tokio::spawn(async move {
       loop {
@@ -222,14 +222,14 @@ fn calculate_next_tick_duration(
   default_interval: Duration,
 ) -> Duration {
   if cfg!(feature = "test_util") {
-    Duration::from_secs(1)
+    Duration::from_millis(500)
   } else if num_messages == 0 {
-    Duration::from_secs(2)
+    Duration::from_secs(1)
   } else {
     match num_init_sync {
       0..=3 => default_interval,
-      4..=7 => Duration::from_secs(3),
-      _ => Duration::from_secs(6),
+      4..=7 => Duration::from_secs(2),
+      _ => Duration::from_secs(4),
     }
   }
 }
